+++ arangod/Aql/CMakeLists.txt
@@ -69,8 +69,6 @@ add_library(arango_aql STATIC
   NodeFinder.cpp
   NonConstExpression.cpp
   NonConstExpressionContainer.cpp
-  Optimizer.cpp
-  OptimizerRules.cpp
   OutputAqlItemRow.cpp
   Parser.cpp
   Projections.cpp
