+++ arangod/Aql/Optimizer/CMakeLists.txt
@@ -2,5 +2,7 @@ add_subdirectory(Rule)
 
 target_sources(arangoserver PRIVATE
   Feature.cpp
+  Optimizer.cpp
   OptionsProvider.cpp
+  Rules.cpp
   Utils.cpp)
