+++ arangod/Aql/Optimizer/Feature.cpp
@@ -28,7 +28,7 @@
 #include "Aql/Optimizer/Rule/Graph.h"
 #include "Aql/Optimizer/Rule/IResearchView.h"
 #include "Aql/Optimizer/Rule/IndexNode.h"
-#include "Aql/OptimizerRules.h"
+#include "Aql/Optimizer/Rules.h"
 #include "Aql/Optimizer/OptionsProvider.h"
 #include "Basics/Exceptions.h"
 #include "Cluster/ServerState.h"
