+++ arangod/Aql/Optimizer/Feature.h
@@ -28,7 +28,7 @@
 #include <vector>
 
 #include "ApplicationFeatures/ApplicationFeature.h"
-#include "Aql/OptimizerRule.h"
+#include "Aql/Optimizer/Rule.h"
 #include "Aql/Optimizer/Options.h"
 
 namespace arangodb {
