+++ arangod/Aql/Optimizer/Rule.h
