+++ arangod/Aql/Optimizer/Rule/Cluster.cpp
@@ -37,9 +37,9 @@
 #include "Aql/ExecutionNode/UpdateReplaceNode.h"
 #include "Aql/ExecutionPlan.h"
 #include "Aql/Expression.h"
-#include "Aql/Optimizer.h"
-#include "Aql/OptimizerRule.h"
-#include "Aql/OptimizerRules.h"
+#include "Aql/Optimizer/Optimizer.h"
+#include "Aql/Optimizer/Rule.h"
+#include "Aql/Optimizer/Rules.h"
 #include "Aql/QueryContext.h"
 #include "Basics/StaticStrings.h"
 #include "Indexes/Index.h"
