+++ arangod/Aql/Optimizer/Rule/CollectWithIndex.cpp
@@ -32,8 +32,8 @@
 #include "Aql/Expression.h"
 #include "Aql/IndexDistrinctScan.h"
 #include "Aql/IndexStreamIterator.h"
-#include "Aql/Optimizer.h"
-#include "Aql/OptimizerRules.h"
+#include "Aql/Optimizer/Optimizer.h"
+#include "Aql/Optimizer/Rules.h"
 #include "Aql/Query.h"
 #include "Indexes/Index.h"
 #include "Logger/LogMacros.h"
