+++ arangod/Aql/Optimizer/Rule/IResearchView.cpp
@@ -41,9 +41,9 @@
 #include "Aql/ExecutionPlan.h"
 #include "Aql/Expression.h"
 #include "Aql/Function.h"
-#include "Aql/Optimizer.h"
+#include "Aql/Optimizer/Optimizer.h"
 #include "Aql/Optimizer/Rule/IResearchView.h"
-#include "Aql/OptimizerRule.h"
+#include "Aql/Optimizer/Rule.h"
 #include "Aql/Projections.h"
 #include "Aql/Query.h"
 #include "Aql/SortCondition.h"
