+++ arangod/Aql/Optimizer/Rule/PropagateConstantAttributes.cpp
@@ -28,8 +28,8 @@
 #include "Aql/ExecutionNode/FilterNode.h"
 #include "Aql/ExecutionNode/SubqueryNode.h"
 #include "Aql/Expression.h"
-#include "Aql/Optimizer.h"
-#include "Aql/OptimizerRules.h"
+#include "Aql/Optimizer/Optimizer.h"
+#include "Aql/Optimizer/Rules.h"
 #include "Aql/Optimizer/Utils.h"
 #include "Cluster/ServerState.h"
 #include "Containers/ImmutableMap.h"
