+++ arangod/Aql/Optimizer/Rule/PushDownLateMaterialization.cpp
@@ -31,9 +31,9 @@
 #include "Aql/ExecutionNode/SubqueryNode.h"
 #include "Aql/ExecutionPlan.h"
 #include "Aql/Expression.h"
-#include "Aql/Optimizer.h"
-#include "Aql/OptimizerRule.h"
-#include "Aql/OptimizerRules.h"
+#include "Aql/Optimizer/Optimizer.h"
+#include "Aql/Optimizer/Rule.h"
+#include "Aql/Optimizer/Rules.h"
 #include "Aql/Optimizer/Utils.h"
 #include "Aql/Query.h"
 #include "Indexes/Index.h"
