+++ arangod/Aql/Optimizer/Rule/PushFilterVectorIndex.cpp
@@ -32,8 +32,8 @@
 #include "Aql/ExecutionNode/LimitNode.h"
 #include "Aql/ExecutionNode/FilterNode.h"
 #include "Aql/ExecutionNode/SortNode.h"
-#include "Aql/Optimizer.h"
-#include "Aql/OptimizerRules.h"
+#include "Aql/Optimizer/Optimizer.h"
+#include "Aql/Optimizer/Rules.h"
 #include "Aql/Optimizer/Utils.h"
 #include "Aql/QueryContext.h"
 #include "Basics/Exceptions.h"
