+++ arangod/Aql/Optimizer/Rule/PushLimitIntoIndex.cpp
@@ -28,8 +28,8 @@
 #include "Aql/ExecutionNode/IndexNode.h"
 #include "Aql/ExecutionNode/LimitNode.h"
 #include "Aql/ExecutionNode/SortNode.h"
-#include "Aql/Optimizer.h"
-#include "Aql/OptimizerRules.h"
+#include "Aql/Optimizer/Optimizer.h"
+#include "Aql/Optimizer/Rules.h"
 #include "Aql/Optimizer/Utils.h"
 #include "Aql/SortElement.h"
 #include "Aql/Variable.h"
