+++ arangod/Aql/Optimizer/Rule/ReplaceEntriesWithObjectIteration.cpp
@@ -29,8 +29,8 @@
 #include "Aql/ExecutionPlan.h"
 #include "Aql/Expression.h"
 #include "Aql/Function.h"
-#include "Aql/Optimizer.h"
-#include "Aql/OptimizerRules.h"
+#include "Aql/Optimizer/Optimizer.h"
+#include "Aql/Optimizer/Rules.h"
 #include "Aql/Optimizer/Utils.h"
 #include "Logger/LogMacros.h"
