+++ arangod/Aql/Optimizer/Rule/ReplaceEqualAttributeAccess.cpp
@@ -36,8 +36,8 @@
 #include "Aql/ExecutionNode/SubqueryNode.h"
 #include "Aql/ExecutionPlan.h"
 #include "Aql/Expression.h"
-#include "Aql/Optimizer.h"
-#include "Aql/OptimizerRules.h"
+#include "Aql/Optimizer/Optimizer.h"
+#include "Aql/Optimizer/Rules.h"
 #include "Aql/Optimizer/Utils.h"
 #include "Aql/Query.h"
 #include "Aql/Variable.h"
