+++ arangod/Aql/Optimizer/Rule/ReplaceFunctions.cpp
@@ -40,8 +40,8 @@
 #include "Aql/Expression.h"
 #include "Aql/Function.h"
 #include "Aql/IndexHint.h"
-#include "Aql/Optimizer.h"
-#include "Aql/OptimizerRules.h"
+#include "Aql/Optimizer/Optimizer.h"
+#include "Aql/Optimizer/Rules.h"
 #include "Aql/Query.h"
 #include "Aql/SortElement.h"
 #include "Aql/Variable.h"
