+++ arangod/Aql/Optimizer/Rule/VectorIndex.cpp
@@ -34,9 +34,9 @@
 #include "Aql/ExecutionNode/LimitNode.h"
 #include "Aql/ExecutionNode/FilterNode.h"
 #include "Aql/ExecutionNode/SortNode.h"
-#include "Aql/Optimizer.h"
-#include "Aql/OptimizerRule.h"
-#include "Aql/OptimizerRules.h"
+#include "Aql/Optimizer/Optimizer.h"
+#include "Aql/Optimizer/Rule.h"
+#include "Aql/Optimizer/Rules.h"
 #include "Aql/Optimizer/Utils.h"
 #include "Aql/Query.h"
 #include "Aql/types.h"
