+++ arangod/Aql/StandaloneCalculation.cpp
@@ -29,8 +29,8 @@
 #include "Aql/ExpressionContext.h"
 #include "Aql/Expression.h"
 #include "Aql/LazyConditions.h"
-#include "Aql/Optimizer.h"
-#include "Aql/OptimizerRule.h"
+#include "Aql/Optimizer/Optimizer.h"
+#include "Aql/Optimizer/Rule.h"
 #include "Aql/Parser.h"
 #include "Aql/QueryContext.h"
 #include "Aql/QueryString.h"
