+++ arangod/IResearch/IResearchAqlAnalyzer.cpp
@@ -32,8 +32,8 @@
 #include "Aql/ExecutionPlan.h"
 #include "Aql/Expression.h"
 #include "Aql/FixedVarExpressionContext.h"
-#include "Aql/Optimizer.h"
-#include "Aql/OptimizerRule.h"
+#include "Aql/Optimizer/Optimizer.h"
+#include "Aql/Optimizer/Rule.h"
 #include "Aql/Parser.h"
 #include "Aql/QueryContext.h"
 #include "Aql/QueryString.h"
