+++ arangod/RocksDBEngine/RocksDBOptimizerRules.cpp
@@ -37,8 +37,8 @@
 #include "Aql/Expression.h"
 #include "Aql/Function.h"
 #include "Aql/IndexHint.h"
-#include "Aql/Optimizer.h"
-#include "Aql/OptimizerRule.h"
+#include "Aql/Optimizer/Optimizer.h"
+#include "Aql/Optimizer/Rule.h"
 #include "Aql/Optimizer/Feature.h"
 #include "Aql/Optimizer/Utils.h"
 #include "Aql/Query.h"
