+++ tests/IResearch/IResearchQueryOptionsTest.cpp
@@ -26,7 +26,7 @@
 
 #include <velocypack/Iterator.h>
 
-#include "Aql/OptimizerRule.h"
+#include "Aql/Optimizer/Rule.h"
 #include "Aql/Query.h"
 #include "IResearch/IResearchLink.h"
 #include "IResearch/IResearchLinkHelper.h"
